Phase 2/Machine Learning/requirements.txt
tensorflow
numpy
pandas
flask
scikit-learn
matplotlib
transformers
tf-keras
nltk
